requirements.txt
annotated-types==0.6.0
ansicon==1.89.0
anyio==4.3.0
arrow==1.3.0
art==6.1
Babel==2.14.0
binaryornot==0.4.4
blessed==1.20.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cookiecutter==2.6.0
cycler==0.12.1
decorator==4.4.2
docxcompose==1.4.0
docxtpl==0.16.8
et-xmlfile==1.1.0
Faker==24.3.0
fastapi==0.110.0
fillpdf==0.7.2
flet==0.21.2
flet-core==0.21.2
flet-runtime==0.21.2
fonttools==4.50.0
gpustat==1.1.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
Jinja2==3.1.3
jinxed==1.2.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
moviepy==1.0.3
numpy==1.26.4
nvidia-ml-py==12.535.133
oauthlib==3.2.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pdf2image==1.17.0
pdfrw2==0.5.0
pillow==10.2.0
proglog==0.1.10
psutil==5.9.8
pyasn1==0.6.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
pypdf==4.1.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
repath==0.9.0
requests==2.31.0
rich==13.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
text-unidecode==1.3
tqdm==4.66.2
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
